Refact : onBoarding Component 폴더로 이전
