+++ src/pages/onBoarding/index.tsx
@@ -1,11 +1,11 @@
 import useOnboardingStore from '@/store/validationStore';
-import KakaoIdPage from '@/pages/onBoarding/KakaoId';
-import LocationPage from '@/pages/onBoarding/Location';
-import SexPage from '@/pages/onBoarding/Sex';
-import NickNamePage from '@/pages/onBoarding/NickName';
-import BirthPage from '@/pages/onBoarding/Birth';
-import ProfileImagePage from '@/pages/onBoarding/profileImage';
-import OneLinerPage from '@/pages/onBoarding/OneLiner';
+import KakaoIdPage from '@/components/onBoarding/KakaoId';
+import LocationPage from '@/components/onBoarding/Location';
+import SexPage from '@/components/onBoarding/Sex';
+import NickNamePage from '@/components/onBoarding/NickName';
+import BirthPage from '@/components/onBoarding/Birth';
+import ProfileImagePage from '@/components/onBoarding/profileImage';
+import OneLinerPage from '@/components/onBoarding/OneLiner';
 
 const OnBoardingPage = () => {
   const { currentPage } = useOnboardingStore();
